chore: remove unused deps, extract getMondayOfDate and formatWeekLabelShort to lib/formatters
