+++ package.json
@@ -17,12 +17,10 @@
   },
   "dependencies": {
     "@base-ui/react": "^1.2.0",
-    "@hookform/resolvers": "^5.2.2",
     "bcryptjs": "^3.0.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
-    "date-fns": "^4.1.0",
     "drizzle-orm": "^0.45.1",
     "embla-carousel-react": "^8.6.0",
     "exceljs": "^4.4.0",
