+++ src/components/notifications/ClosedCasesTab.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, useRef, Fragment } from "react";
 import { ChevronLeft, ChevronRight, RefreshCw, CheckCircle, AlertCircle, ExternalLink } from "lucide-react";
 import { themeClasses } from "@/lib/theme-classes";
-import { getMonday, formatWeekLabel as formatWeekLabelBase } from "@/lib/formatters";
+import { getMonday, formatWeekLabelShort as formatWeekLabel } from "@/lib/formatters";
 
 interface DayCount {
   date: string;
@@ -23,7 +23,6 @@ interface ClosedCasesTabProps {
   t: ReturnType<typeof themeClasses>;
 }
 
-const formatWeekLabel = (monday: string): string => formatWeekLabelBase(monday, 6);
 
 const fmtDate = (iso: string): string =>
   new Date(`${iso}T00:00:00`).toLocaleDateString("en-US", {
